NRI: added resolve functionality: - added "CmdResolveTexture" (for MSAA resolve at least) - extended "StageBits", "AccessBits" and "Layout" - relaxed valid usage of "[src/dst]RegionDesc", which can be 0/non-0 independently

+++ Source/Validation/CommandQueueVal.hpp
@@ -13,7 +13,7 @@ static bool ValidateTextureUploadDesc(DeviceVal& device, uint32_t i, const Textu
     RETURN_ON_FAILURE(&device, textureUploadDesc.after.layout < Layout::MAX_NUM, false, "'textureUploadDescs[%u].nextLayout' is invalid", i);
     RETURN_ON_FAILURE(&device, textureVal.IsBoundToMemory(), false, "'textureUploadDescs[%u].texture' is not bound to memory", i);
 
-    uint32_t subresourceNum = textureDesc.layerNum * textureDesc.mipNum;
+    uint32_t subresourceNum = (uint32_t)textureDesc.layerNum * (uint32_t)textureDesc.mipNum;
     for (uint32_t j = 0; j < subresourceNum; j++) {
         const TextureSubresourceUploadDesc& subresource = textureUploadDesc.subresources[j];
